refactor: move VFileSystem to its own file.
