+++ packages/cspell-io/src/VirtualFS/redirectProvider.ts
@@ -2,7 +2,8 @@ import assert from 'node:assert';
 
 import { renameFileReference, renameFileResource, urlOrReferenceToUrl } from '../common/index.js';
 import type { DirEntry, FileReference, FileResource } from '../models/index.js';
-import type { FSCapabilityFlags, VFileSystemProvider, VProviderFileSystem } from '../VirtualFS.js';
+import type { FSCapabilityFlags } from '../VFileSystem.js';
+import type { VFileSystemProvider, VProviderFileSystem } from '../VirtualFS.js';
 import { fsCapabilities, VFSErrorUnsupportedRequest } from '../VirtualFS.js';
 
 type UrlOrReference = URL | FileReference;
