+++ packages/cspell-io/src/VirtualFs.test.ts
@@ -7,8 +7,9 @@ import { CFileResource } from './common/index.js';
 import { FileType } from './models/Stats.js';
 import { toFileURL, urlBasename } from './node/file/url.js';
 import { pathToSample as ps } from './test/test.helper.js';
+import { FSCapabilityFlags } from './VFileSystem.js';
 import type { VFileSystemProvider, VirtualFS, VProviderFileSystem } from './VirtualFS.js';
-import { createVirtualFS, FSCapabilityFlags, getDefaultVirtualFs, VFSErrorUnsupportedRequest } from './VirtualFS.js';
+import { createVirtualFS, getDefaultVirtualFs, VFSErrorUnsupportedRequest } from './VirtualFS.js';
 
 const sc = expect.stringContaining;
 const oc = expect.objectContaining;
