+++ packages/cspell-io/src/index.ts
@@ -17,13 +17,8 @@ export type { Stats } from './models/Stats.js';
 export { FileType as VFileType } from './models/Stats.js';
 export { encodeDataUrl, toDataUrl } from './node/dataUrl.js';
 export { isFileURL, isUrlLike, toFileURL, toURL, urlBasename, urlDirname } from './node/file/url.js';
-export type {
-    VFileSystem as VFileSystem,
-    VFileSystemProvider,
-    VfsDirEntry,
-    VfsStat,
-    VirtualFS,
-    VProviderFileSystem,
-} from './VirtualFS.js';
-export { createVirtualFS, FSCapabilityFlags, getDefaultVirtualFs } from './VirtualFS.js';
+export type { VFileSystem, VfsDirEntry, VfsStat } from './VFileSystem.js';
+export { FSCapabilityFlags } from './VFileSystem.js';
+export type { VFileSystemProvider, VirtualFS, VProviderFileSystem } from './VirtualFS.js';
+export { createVirtualFS, getDefaultVirtualFs } from './VirtualFS.js';
 export { createRedirectProvider } from './VirtualFS/redirectProvider.js';
